Address review: abstract web fullscreen import, drop lint ignore
- Remove `// ignore: uri_does_not_exist`; chewie_player.dart now imports
  the plain `web_fullscreen.dart` abstraction, which conditionally
  exports the stub or the real implementation.
- Use `dart.library.js_interop` instead of `dart.library.html` for the
  conditional export.
- Move the dart:js_interop / package:web implementation into
  web_fullscreen_impl.dart.
